==> Mapping/01_FooDB_FooDBCleaning.ipynb <==
# Database Matching: FooDB to ASA24 Ingredient Descriptions
Stephanie Wilson, January 2023

### Step 1: Clean FooDB's 'Content.csv'

__Required Input Files__

  - **Content.csv.bz2** - version 1, downloaded direct from foodb.ca
  - **Food.csv** - version 1, downloaded direct from foodb.ca  
 

__Information__  
This script reformats and fills in missing food descriptions in FoodB's Content.csv. Specifically, this code:
    
   1) Removes items derived from Pathbank (Lipid from Arabidopsis) and HMDB  
   2) Identifies missing common food descriptions  
   3) Add food descriptions from Food.csv
       - This can lose additional information about preparation. However, this is the best descriptor possible given the information provided.
   4) Identify which food descriptions are still missing and find closest description match.
       - Output: Food_updated.csv
   5) Update Content file to include food descriptions
   6) Export updated Content file
       - Output: Content_updated.csv
       
__Outputs__

  - **Content_updated.csv.bz2** - Food descriptions were added to all rows missing food descriptions in the original
  - **Food_updated.csv** - Food 554 was added to the original Food.csv

#Load modules
import os
import pandas as pd
import numpy as np
import re
import bz2

#Ensure working directory is the project folder
mapping = os.getcwd()
mapping

#Load data
Content = pd.read_csv('FooDB/Content.csv.bz2', compression='bz2', low_memory=False)
Food = pd.read_csv('FooDB/Food.csv')

### 1) Remove rows from HMDB and Pathpank

We are looking for compounds specifically within foods. So we will remove: 
  - endogenously produced compounds from humans (HMDB)
  - lipids from arabidopsis (Pathbank)
  
Rows meeting these criteria were all predicted and did NOT contain any content values. 

#Remove data from Pathbank and HMDB
Content = Content[Content["orig_food_common_name"]!= 'Endogenous compounds from human (HMDB)']
Content = Content[Content['orig_food_common_name']!='Lipid from Arabidopsis (PathBank)']
Content.shape
#cuts numbers of rows down drastically

### 2) Identify which rows do not have food descriptions

#Pull out rows with missing common food descriptions
missing = Content[Content['orig_food_common_name'].isnull()]
print(missing.shape[0], ' rows are missing common food descriptions')

### 3) Add food descriptions from Food.csv

# Rename in Food.csv to match column in Content
Food = Food.rename(columns = {'id':'food_id'})

# Filter in specific rows
Food = Food.filter(['food_id', 'name'], axis = 1)

# Merge missing data with food.csv
missing = pd.merge(missing, Food, on = 'food_id', how = 'left')
print(missing.shape[0], ' rows are missing common food descriptions')

# Replace orig_food_common_name (all missing) with matched name from Food.csv
missing['orig_food_common_name'] = missing['name']

# Drop the name column as we've transferred the names over to another column
missing.drop(columns=['name'], inplace=True)

### 4) Identify which food descriptions are still missing and rectify.

# What didn't get a name?
still_missing = missing[missing['orig_food_common_name'].isnull()]
print(still_missing.shape[0], 'rows are still missing common food descriptions')

# What food codes did not have a food description?
still_missing['food_id'].value_counts()
# food_id codes 554 and 16420 did not get a common name code

#### 4a) Resolve food_id 554

# What information do we have in FooDB Content on #554?
id554 = Content[(Content['food_id']==554)]
id554['orig_food_common_name'].value_counts()
# We can see that food_id 554 is equivalent to 'Owl, horned, flesh, raw (Alaska Native)'

Food.loc[(Food['food_id']==554)]
# Food 554 does not exist in Food.csv. Add this entry

# What information do we have in FooDB Content on #16420?
id16420 = Content[(Content['food_id']==16420)]
id16420
# We can see that food_id 554 is equivalent to 'Soy protein concentrate, aqueous washed'

# Create Entry for 554, matches Food.csv col names
id554_entry = {'id':[554], 'name':['Owl, horned, flesh, raw (Alaska Native)']} 
id554_entry = pd.DataFrame(id554_entry)

# Reload Food.csv to make edits
Food_orig = pd.read_csv('FooDB/Food.csv')

# Add 554 to Foodb_food aka Food.csv
Food_orig = pd.concat([Food_orig, id554_entry], ignore_index = True)
Food_orig.to_csv('FooDB/Food_updated.csv', index = None, header = True)

#### 4b) Resolve food_id 16420

# Check if the code is a remnant from the database the compound was derived from and just didn't get converted.
id16420_orig = Content[(Content['orig_food_id']=='16420')]
id16420_orig
# Confirmed Code 16420 corresponds to FooDB's food codes 272 and 85

# What is the usage like for unique food codes 272 and 85?
id16420_orig['food_id'].value_counts()

# What are the food descriptors for 272 and 85
Food.loc[(Food['food_id']==272)|(Food['food_id']==85)]
# Will choose to classify id 16420 with food_id 272 'Other soy product' moving forward as it is more similar to FNDDS usage

still_missing
# Next steps:
# Replace 554 name with 'Owl, horned, flesh, raw (Alaska Native)'
# Replace 16420 id with 272 and name with 'Soy protein concentrate, aqueous washed'

# Replace 554 name with 'Owl, horned, flesh, raw (Alaska Native)'
still_missing.loc[still_missing['food_id']==554, 'orig_food_common_name'] = 'Owl, horned, flesh, raw (Alaska Native)'
still_missing

# Replace 16420 name description with 'Soy protein concentrate, aqueous washed'
still_missing.loc[still_missing['food_id']==16420, 'orig_food_common_name'] = 'Soy protein concentrate, aqueous washed'
still_missing

# Replace 16420 with 272 
still_missing['food_id'].replace(to_replace = 16420, value = 272, inplace=True)
still_missing

### 5) Update Content file

Recall: missing was updated to include food descriptions from Food.csv. Anything that didn't have a description went on to still_missing for identification.

We will now drop those rows with a missing description and add still_missing which has the added descriptions. 

# drop rows wih missing description
missing = missing.dropna(subset=['orig_food_common_name'])

# add still missing (updated now with names!) to missing_
missing_updated = pd.concat([missing, still_missing], ignore_index = True)

# Confirm there are now no missing food descriptions
missing_check = missing_updated[missing_updated['orig_food_common_name'].isnull()]
print(missing_check.shape[0], ' rows are missing common food descriptions')

We will repeat the above process to fill in the Content file.

# drop rows wih missing description
Content = Content.dropna(subset=['orig_food_common_name'])

# add missing_updated to Content
Content_updated = pd.concat([Content, missing_updated], ignore_index = True)

# Confirm there are now no missing food descriptions
Content_check = Content_updated[Content_updated['orig_food_common_name'].isnull()]
print(Content_check.shape[0], ' rows are missing common food descriptions')

# Establish a unique number for each group of food names, starting from 1
Content_updated['food_V2_ID'] = pd.factorize(Content_updated['orig_food_common_name'])[0] + 1

### 6) Export Updated Content File

Export the content file now updated and complete with food descriptions for every entry.

Content_updated.to_csv('FooDB/Content_updated.csv.bz2', index = None, header = True)

#Matches the original dimensions + 1 for the duplicate ID column
Content_updated.shape
